unify retrieval names
