+++ torchmetrics/retrieval/r_precision.py
@@ -11,16 +11,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Any, Callable, Optional
-
 from torch import Tensor, tensor
 
-from torchmetrics.functional.retrieval.recall import retrieval_recall
-from torchmetrics.retrieval.retrieval_metric import RetrievalMetric
+from torchmetrics.functional.retrieval.r_precision import retrieval_r_precision
+from torchmetrics.retrieval.metric_base import RetrievalMetric
 
 
-class RetrievalRecall(RetrievalMetric):
-    """Computes `IR Recall`_.
+class RetrievalRPrecision(RetrievalMetric):
+    """Computes `IR R-Precision`_.
 
     Works with binary target data. Accepts float predictions from a model output.
 
@@ -32,8 +30,8 @@ class RetrievalRecall(RetrievalMetric):
 
     ``indexes``, ``preds`` and ``target`` must have the same dimension.
     ``indexes`` indicate to which query a prediction belongs.
-    Predictions will be first grouped by ``indexes`` and then `Recall` will be computed as the mean
-    of the `Recall` over each query.
+    Predictions will be first grouped by ``indexes`` and then `R-Precision` will be computed as the mean
+    of the `R-Precision` over each query.
 
     Args:
         empty_target_action:
@@ -46,7 +44,6 @@ class RetrievalRecall(RetrievalMetric):
 
         ignore_index:
             Ignore predictions where the target is equal to this number.
-        k: consider only the top k elements for each query (default: `None`, which considers them all)
         compute_on_step:
             Forward only calls ``update()`` and return None if this is set to False.
         dist_sync_on_step:
@@ -63,43 +60,18 @@ class RetrievalRecall(RetrievalMetric):
             If ``empty_target_action`` is not one of ``error``, ``skip``, ``neg`` or ``pos``.
         ValueError:
             If ``ignore_index`` is not `None` or an integer.
-        ValueError:
-            If ``k`` parameter is not `None` or an integer larger than 0.
 
     Example:
-        >>> from torchmetrics import RetrievalRecall
+        >>> from torchmetrics import RetrievalRPrecision
         >>> indexes = tensor([0, 0, 0, 1, 1, 1, 1])
         >>> preds = tensor([0.2, 0.3, 0.5, 0.1, 0.3, 0.5, 0.2])
         >>> target = tensor([False, False, True, False, True, False, True])
-        >>> r2 = RetrievalRecall(k=2)
-        >>> r2(preds, target, indexes=indexes)
+        >>> p2 = RetrievalRPrecision()
+        >>> p2(preds, target, indexes=indexes)
         tensor(0.7500)
     """
 
     higher_is_better = True
 
-    def __init__(
-        self,
-        empty_target_action: str = "neg",
-        ignore_index: Optional[int] = None,
-        k: Optional[int] = None,
-        compute_on_step: bool = True,
-        dist_sync_on_step: bool = False,
-        process_group: Optional[Any] = None,
-        dist_sync_fn: Callable = None,
-    ) -> None:
-        super().__init__(
-            empty_target_action=empty_target_action,
-            ignore_index=ignore_index,
-            compute_on_step=compute_on_step,
-            dist_sync_on_step=dist_sync_on_step,
-            process_group=process_group,
-            dist_sync_fn=dist_sync_fn,
-        )
-
-        if (k is not None) and not (isinstance(k, int) and k > 0):
-            raise ValueError("`k` has to be a positive integer or None")
-        self.k = k
-
     def _metric(self, preds: Tensor, target: Tensor) -> Tensor:
-        return retrieval_recall(preds, target, k=self.k)
+        return retrieval_r_precision(preds, target)
